# services/api/openapi.yaml
openapi: 3.0.3
info:
  title: ReceiptRadar API
  version: 0.1.0
servers:
  - url: http://localhost:3000
paths:
...
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
...
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user:
          $ref: '#/components/schemas/UserProfile'
    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        city:
          type: string
        consents:
          type: object
...